python/requirements.txt
fastapi 
uvicorn
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi